Controller, harness and plant working
In the sh_plant.slx file there is a model of the plant considered for the Skyhook controller. In particular the plant refers to a quarter car model. The sh_controller.slx file contain the controller and its interfaces, in particular this file contains all the controller logic to produce the outputs and the error, developed mainly in Stateflow. The sh_harness.slx contains an harness to test both plant and controller. The .m file contains some initializations and a first test using low horizontal speed of the car and a relative smooth road profile.